requirements.txt
streamlit
fastapi
uvicorn
python-dotenv
requests
beautifulsoup4
lxml
openai
google-generativeai
groq
transformers
torch